working build. no usart yet :(

+++ run.sh
@@ -1,1 +1,1 @@
-make flash && rm -f main.o
+clear && make flash && rm -f main.o
